Refonte des modèles et affichage dynamique du formulaire d'analyse du DPO
