+++ app_edcp/enregistrement/templates/enregistrement/enregistrement_detail.html
@@ -10,7 +10,7 @@
   {% comment %}   <div class="col-3"></div> {% endcomment %}
   <div>
     <div class="col">
-      <h3 class="text-hint mb-3 custom-text-light fw-bold">Détails de l'organisation</h3>
+      <h3 class="mb-3 custom-text-light fw-bold">Détails de l'organisation</h3>
     </div>
   </div>
